src/Producer/requirements.txt
pandas
kafka-python
redis
